+++ Joker.Kafka/Infrastructure/Extensions/StringExtensions.cs
@@ -1,0 +1,19 @@
+﻿using System.Linq;
+using System.Text.RegularExpressions;
+
+namespace Kafka.DotNet.ksqlDB.Infrastructure.Extensions
+{
+  internal static class StringExtensions
+  {
+    public static string ToKSqlFunctionName(this string functionName)
+    {      
+      var words = 
+        Regex.Matches(functionName, @"([A-Z][a-z]+)")
+          .Select(m => m.Value);
+
+      var ksqlFunctionName = string.Join("_", words).ToUpper();
+
+      return ksqlFunctionName.ToUpper();
+    }
+  }
+}